Organized scripts into different toolsets
