+++ CollectorUtils/add_update_gnss_fields_python_api.md
