+++ CollectorUtils/configure_gnss_popup_python_api.md
